Move dynamic layout components to subdirectory
