+++ components/Template.tsx
@@ -1,9 +1,9 @@
 import * as React from "react"
+import styled from "styled-components"
 import { urlFor, usePath, Development, StyledSheet, useContext } from "monobase"
 import { FramerAPIDefaultProvider } from "./contexts/FramerAPIContext"
 import { Sidebar } from "./layout/Sidebar"
 import { Codebar } from "./layout/Codebar"
-import styled from "styled-components"
 import { tablet, mobile } from "./layout/Breakpoints"
 import { Markdown } from "./layout/Markdown"
 import { GoogleTag } from "./GoogleTag"
